drf-auth-and-validation/api_with_restrictions/requirements.txt
Django==3.1.2
djangorestframework==3.12.1
psycopg2-binary==2.8.6
django-filter==2.4.0

django-debug-toolbar~=3.2.1